+++ src/frontend/components/property-value/property-value.ts
@@ -9,11 +9,11 @@ import {Highlightable} from '../../utils/highlightable';
 
 @Component({
   selector: 'bt-property-value',
-  template: require('./property-value.html'),
+  templateUrl: './property-value.html',
 })
 export class PropertyValue extends Highlightable {
-  @Input() private key: string;
-  @Input() private value: string;
+  @Input() key: string;
+  @Input() value: string;
 
   constructor(@Inject(ChangeDetectorRef) changeDetectorRef: ChangeDetectorRef) {
     super(changeDetectorRef, changes => changes && changes.hasOwnProperty('value'));